12/01bis

+++ srcs/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: vmusunga <vmusunga@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/12 11:25:11 by vmusunga          #+#    #+#             */
-/*   Updated: 2021/01/12 11:25:53 by vmusunga         ###   ########.fr       */
+/*   Updated: 2021/01/12 19:00:49 by vmusunga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
