+++ srcs/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: vmusunga <vmusunga@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/12 11:26:23 by vmusunga          #+#    #+#             */
-/*   Updated: 2021/01/12 11:27:57 by vmusunga         ###   ########.fr       */
+/*   Updated: 2021/01/12 19:01:38 by vmusunga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int i;
+	size_t i;
 		
 	i = 0;
 	if (n == 0)
